refactor(core): split observation/store god-file into module dir
Splits src/core/observation/store.rs (1510 lines \u2014 baselined GodFile) into a
store/ module dir (structs + re-exports in mod.rs, ObservationStore impl split
across artifacts/runs/findings/triage_items/schema submodules, helpers in
helpers.rs), each under thresholds. Pure move; verified cargo build --lib
--tests (exit 0). GodFile baseline row removed.
